requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.7
python-dotenv>=1.0
openai>=1.99.0,<2.0.0
httpx>=0.27
